add a Socket#monitor method
Sets up an internal pair socket that publishes socket events to an
inproc endpoint. Another pair socket can be connected to the endpoint to
read those events.

Only works with ZeroMQ 3.2+ and ffi-rzmq > 0.9.7. The method is not
defined for earlier versions of ffi-rzmq because it doesn't have the
necessary constants.

See http://api.zeromq.org/3-2:zmq-socket-monitor for more info.